Перенос fill_u32 в Helper.h/cpp.
